+++ NEWS.md
@@ -1,5 +1,13 @@
 # esqlabsR 3.0.0.9000
 
+NEW FUNCTIONS
+
+- To carry out and visualize sensitivity analysis:
+* `sensitivityCalculation()`
+* `sensitivitySpiderPlot()`
+* `sensivitityTimeProfiles()`
+* `savePlotList()`
+
 BREAKING CHANGES
 
 - The package gains a new dependency: [`{ospsuite.utils}`](https://www.open-systems-pharmacology.org/OSPSuite.RUtils/).